Change how the automaton represents actions, and add logging statements
There was an error with the way that the actions for the rover were
handled. When sending multiple commands they would overwrite each other
and the rover might not execute the right action. By making the actions
explicit and exclusive, only a single action will be accomplished at a
time.