Add sentinel value to rover properties to ensure they are set when switched
If you set the velocity of the rover, then the angular velocity, then
the velocity to the initial value, the rover would not stop turning
because the velocity was not reset when the angular velocity was set.
Using this sentinel value will allow the rover to know when its modes
are switched so it can make sure to send the correct commands.